requirements.txt
streamlit==1.32.0
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
scikit-learn==1.4.1.post1
